Move render_grpc_request to yaak-models too; drop yaak::render
